Refactor separating functionality into better namespaces
